+++ types/wdio-intercept-service.d.ts
@@ -16,43 +16,85 @@ declare namespace WdioInterceptorService {
     statusCode: number;
   }
 
-  interface InterceptedRequest {
+  interface PendingRequest {
     url: string;
     method: HTTPMethod;
-    body: string | object;
+    body: string | null | object;
+    headers: object;
+    pending: true;
+  }
+
+  interface CompletedRequest {
+    url: string;
+    method: HTTPMethod;
+    body: string | null | object;
+    pending: false;
     headers: object;
     response: {
       headers: object;
-      body: string | object;
+      body: string | null | object;
       statusCode: number;
     };
   }
+
+  type InterceptedRequest = PendingRequest | CompletedRequest;
+
+  interface OrderingStrategy {
+    /** Whether requests are ordered by time of initiation, or fulfillment */
+    orderBy?: 'START' | 'END';
+  }
+  interface AssertionRequestOrderStrategy extends OrderingStrategy {
+    /** Whether the intercepted request order must match the order in which expectations were set */
+    inOrder?: boolean;
+  }
+
+  interface GetRequestOptions extends OrderingStrategy {
+    /** Whether pending requests will be included in the response */
+    includePending?: boolean;
+  }
+  type OnlyCompletedRequests = { includePending: false };
 }
 /**
  * Convert T to T or Promise<T> depending if `@wdio/sync` is being used or not.
  */
-type AsyncSync<T> = WebdriverIO.BrowserObject extends WebDriver.Client
+type AsyncSync<T> = WebdriverIO.Browser extends WebDriver.Client
   ? T
   : Promise<T>;
 
 declare module WebdriverIO {
   interface Browser {
-    setupInterceptor: () => AsyncSync<void>;
-    expectRequest: (
+    setupInterceptor(): AsyncSync<void>;
+    expectRequest(
       method: WdioInterceptorService.HTTPMethod,
       url: string | RegExp,
       statusCode: number
-    ) => AsyncSync<BrowserObject>;
-    assertRequests: () => AsyncSync<BrowserObject>;
-    assertExpectedRequestsOnly: (inOrder?: boolean) => AsyncSync<BrowserObject>;
-    resetExpectations: () => AsyncSync<BrowserObject>;
-    getExpectations: () => AsyncSync<
-      WdioInterceptorService.ExpectedRequest[]
-    >;
-    getRequest: (
-      index: number
-    ) => AsyncSync<WdioInterceptorService.InterceptedRequest>;
-    getRequests: () => AsyncSync<WdioInterceptorService.InterceptedRequest[]>;
+    ): AsyncSync<Browser>;
+    hasPendingRequests(): AsyncSync<boolean>;
+    assertRequests(
+      options?: WdioInterceptorService.OrderingStrategy
+    ): AsyncSync<Browser>;
+    assertExpectedRequestsOnly(inOrder?: boolean): AsyncSync<Browser>;
+    assertExpectedRequestsOnly(
+      options?: WdioInterceptorService.AssertionRequestOrderStrategy
+    ): AsyncSync<Browser>;
+    resetExpectations(): AsyncSync<Browser>;
+    getExpectations(): AsyncSync<WdioInterceptorService.ExpectedRequest[]>;
+    getRequest(
+      index: number,
+      options?: WdioInterceptorService.GetRequestOptions &
+        WdioInterceptorService.OnlyCompletedRequests
+    ): AsyncSync<WdioInterceptorService.CompletedRequest>;
+    getRequest(
+      index: number,
+      options: WdioInterceptorService.GetRequestOptions
+    ): AsyncSync<WdioInterceptorService.InterceptedRequest>;
+    getRequests(
+      options?: WdioInterceptorService.GetRequestOptions &
+        WdioInterceptorService.OnlyCompletedRequests
+    ): AsyncSync<WdioInterceptorService.CompletedRequest[]>;
+    getRequests(
+      options: WdioInterceptorService.GetRequestOptions
+    ): AsyncSync<WdioInterceptorService.InterceptedRequest[]>;
   }
 }